Move current place offer logic to "PlaceOfferLite"
