+++ src/pages/LendPage/components/PlaceOfferTab/PlaceLiteOffer/PlaceLiteOffer.module.less
@@ -1,0 +1,85 @@
+.content {
+  padding: 24px;
+
+  @media (max-width: 640px) {
+    padding: 16px;
+  }
+}
+
+.offerHeaderContent {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 8px;
+
+  @media (max-width: 640px) {
+    margin-bottom: 0;
+  }
+}
+
+.title {
+  font: var(--title-sm);
+  text-transform: uppercase;
+  line-height: 38px;
+
+  @media (max-width: 640px) {
+    font: var(--title-xs);
+    line-height: 32px;
+  }
+}
+
+.offerSummary {
+  border: 1px dotted var(--bg-border);
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 16px;
+  margin: 24px 0;
+
+  @media (max-width: 640px) {
+    gap: 4px;
+    margin: 16px 0;
+  }
+}
+
+.fields {
+  display: flex;
+  align-items: center;
+  gap: 24px;
+  width: 100%;
+
+  @media (max-width: 640px) {
+    gap: 16px;
+  }
+}
+
+.numericField {
+  width: 100%;
+}
+
+.buttonsWrapper {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
+
+.button {
+  height: 40px;
+  width: 100%;
+}
+
+.deleteOfferButton {
+  border: 1px solid var(--additional-red-primary) !important;
+
+  &:hover {
+    border: 1px solid var(--additional-red-primary-deep) !important;
+  }
+}
+
+.messageContainer {
+  height: 14px;
+}
+
+.borrowerMessage {
+  font: var(--body-text-sm);
+}
