+++ src/pages/LendPage/components/PlaceOfferTab/PlaceLiteOffer/helpers.ts
