+++ src/pages/LendPage/components/PlaceOfferTab/PlaceLiteOffer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PlaceLiteOffer'
